Refactor date handling by moving date conversion functions to a utility file for better code organization and reuse
