+++ src/pages/BrewForm.tsx
@@ -4,22 +4,7 @@ import { Bean } from '../types/Bean';
 import { Brew } from '../types/Brew';
 import { useBrewContext } from '../context/BrewContext';
 import StarRating from '../components/StarRating';
-
-const fromUtcToLocalDateTime = (utcDateString: string | undefined) => {
-  if (!utcDateString) return '';
-  const date = new Date(utcDateString);
-  const offset = date.getTimezoneOffset();
-  const localDate = new Date(date.getTime() - offset * 60 * 1000);
-  return localDate.toISOString().slice(0, 16);
-}
-
-const fromLocalToUtc = (localDateString: string | undefined) => {
-  if (!localDateString) return '';
-  const localDate = new Date(localDateString);
-  const offset = localDate.getTimezoneOffset();
-  const date = new Date(localDate.getTime() + offset * 60 * 1000);
-  return date.toISOString();
-}
+import { fromUtcToLocalDateTime, fromLocalToUtc } from '../utils/date';
 
 const BrewForm: React.FC = () => {
   const { beans, brews, updateBrew, setBrews } = useBrewContext();
